投稿の並び順変更

+++ todo/app/Http/Controllers/TaskController.php
@@ -4,12 +4,13 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\TaskRequest;
 use App\Task;
+use Illuminate\Support\Facades\Gate;
 
 class TaskController extends Controller
 {
     public function index()
     {
-        $tasks = Task::all();
+        $tasks = Task::orderBy('created_at', 'desc')->get();
 
         return view('tasks.index', compact('tasks'));
 
@@ -33,11 +34,19 @@ class TaskController extends Controller
 
     public function edit(Task $task)
     {
+        if (Gate::denies('access-task', $task)) {
+            abort(403);
+        }
+
         return view('tasks.edit', compact('task'));
     }
 
     public function update(Task $task, TaskRequest $request)
     {
+        if (Gate::denies('access-task', $task)) {
+            abort(403);
+        }
+
         $task->update([
             'title' => $request->title,
             'contents' => $request->contents,
@@ -48,6 +57,10 @@ class TaskController extends Controller
 
     public function delete(Task $task)
     {
+        if (Gate::denies('access-task', $task)) {
+            abort(403);
+        }
+
         $task->delete();
 
         return redirect()->route('task.index');
